[fe] refactor : query 리팩토링 - 드라이버 매칭페이지와 유저매칭페이지에서 사용하는 gql query들을 `@queries/`로 이동
